docs(source): Rename source index page to `Packages`
Abstract
========

Update `internals.rst` template and apidoc conf to set the name of the
root page of the apidoc part of the doc to "Packages".

Motivation
==========

When going to `source/modules.html` doc page, the title was "isshub",
and also in the breadcum at the top, so we had `isshub > isshub`, which
was not clear.

We needed to change this to a more meaningful name.

Rationale
=========

This is done by setting the `--doc-project` option to `apidoc`, and we
use the same name in `internals.rst`.