docs(source): Lighten source code pages
Abstract
========

Use our own templates for apidoc.

Motivation
==========

Two main reasons:
- "Subpackages" and "Submodules" where present everywhere, in toctree
and pages
- packages/modules always and fully qualified names and as they can be
long, where to heavy for the pages

Rationale
=========

We copied the apidoc templates from
`site-packages/sphinx/templates/apidoc/` to `docs/apidoc_templates/` and
heavily updated their content, also playing with CSS where we cannot
change the rendering (the templates are RST and we don't control how
they are rendered to html), so thanks to the `rst-class` directive we
could play.

In particular:
- remove "Subpackages" and "Submodules" from toctrees
- add them back as a simple <p> in the pages
- use only the non-qualified package name in the title of the page
(lighter toctree in the sidebar and in breadcrumb)
- hide this title and display a little more complete
- and display the "owner package" on the right of the title