docs: Hide "View page source" links
Abstract
========

Apply "display: none" to "View page source" links

Motivation
==========

The "View page source" links point to the source updated by the
`sphinxprettysearchresults` sphinx extension so it was not the real
sources.

In addition, I don't see the point of having access to the RST sources.

Rationale
=========

Instead of trying to remove them from the source, the chosen solution is
a very easy one.