docs(source): Stop duplicating class attributes
Abstract
========

Ask napoleon sphinx extension to put attributes in a "var" section and
not as "attribute" anymore, and hide all "attribute".

Motivation
==========

In the documentation of classes, attributes were duplicated because we
had the ones found by napoleon, with types and docs, then the ones found
by apidoc, without types nor docs.

Rationale
=========

It is not possible to hook into the sphinx process to not display
duplicate attributes because sphinx starts to ask extensions (napoleon)
to do their work, then add its own attributes. We have a hook at this
moment but with no access to the attributes found by napoleon.

So, thanks to the `napoleon_use_ivar` settings, we can ask napoleon to
render the attribute it found in a specific section. We choose to keep
the ones from napoleon instead of the ones found by apidoc itself
because napoleon gives us types and doc.

Pros
====
The attributes are now in a visually separated section so are more
distinguishable from the methods of the class.

Cons
====
If a class attribute is not defined in the class docstring, it won't be
shown at all.