style: Prefix `id_is_positive_integer` with `validate_`
Abstract
========

In `isshub.domain.utils.entity.BaseModelWithId`, rename
`id_is_positive_integer` to `validate_id_is_positive_integer`.

Motivation
==========

Having a meaningful name is the first thing to help knowing what a
function does. Here, without the `validate_` prefix, we couldn't guess
at first sight, in the documentation (where the decorators are not
shown) that it was a validator function.

Rationale
=========

Nothing to do more that renaming the method: the `@field_validator`
track its methods and we don't have to call them directly.