docs(source): Make links between python files in source and git
Abstract
========

- In the files page of the "git" part of the documentation, if they are
still existing python files, add a link to the "source" part of the
documentation.
- In the "source" part of the documentation, add a link to the "git"
part of the documentation, pointing to the file page, with git history.

Motivation
==========

Both view made possible to view the actual code source. But without
links between them, it was not really an option to have:
- in the "git" view, an abstract of the content of a python
package/module
- in the "source" view, an idea of how the file evolved

Rationale
=========

N/A